+++ src/component/Header.tsx
@@ -1,5 +1,6 @@
 import { Container, Nav, Navbar } from 'react-bootstrap';
 import { Link, useLocation } from 'react-router-dom';
+
 export default function Header() {
   const location = useLocation(); // 현재 경로를 가져옴
   return (
